gitstat/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pwdlib
pyjwt
typing